pausing work
I won't be able to work on this for at least 2 weeks- but it's a good
pause point anyway. I think I want to move from the York Lava library to
Kansas Lava.  It has much more type safety and has already tackled many
of the issues that I'm struggling to work with in the York Lava library.